REF: Rename `utils` module to `question`
skipci
